requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
cryptography==46.0.5
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.133.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
iniconfig==2.3.0
oracledb==3.4.2
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
python-dotenv==1.2.1
python-multipart==0.0.27
requests==2.33.1
SQLAlchemy==2.0.47
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
